cleanup some warnings

+++ parallel_hashmap/phmap_bits.h
@@ -279,7 +279,7 @@ PHMAP_BASE_INTERNAL_FORCEINLINE int CountLeadingZeros64(uint64_t n) {
     // MSVC does not have __buitin_clzll. Use _BitScanReverse64.
     unsigned long result = 0;  // NOLINT(runtime/int)
     if (_BitScanReverse64(&result, n)) {
-        return 63 - result;
+        return (int)(63 - result);
     }
     return 64;
 #elif defined(_MSC_VER)
@@ -322,7 +322,7 @@ PHMAP_BASE_INTERNAL_FORCEINLINE int CountLeadingZeros32(uint32_t n) {
 #if defined(_MSC_VER)
     unsigned long result = 0;  // NOLINT(runtime/int)
     if (_BitScanReverse(&result, n)) {
-        return 31 - result;
+        return (int)(31 - result);
     }
     return 32;
 #elif defined(__GNUC__)
@@ -359,7 +359,7 @@ PHMAP_BASE_INTERNAL_FORCEINLINE int CountTrailingZerosNonZero64(uint64_t n) {
 #if defined(_MSC_VER) && defined(_M_X64)
     unsigned long result = 0;  // NOLINT(runtime/int)
     _BitScanForward64(&result, n);
-    return result;
+    return (int)result;
 #elif defined(_MSC_VER)
     unsigned long result = 0;  // NOLINT(runtime/int)
     if (static_cast<uint32_t>(n) == 0) {
@@ -392,7 +392,7 @@ PHMAP_BASE_INTERNAL_FORCEINLINE int CountTrailingZerosNonZero32(uint32_t n) {
 #if defined(_MSC_VER)
     unsigned long result = 0;  // NOLINT(runtime/int)
     _BitScanForward(&result, n);
-    return result;
+    return (int)result;
 #elif defined(__GNUC__)
     static_assert(sizeof(int) == sizeof(n),
                   "__builtin_ctz does not take 32-bit arg");
